Complete the code test
A) Describe the strategy used to consume the API endpoints and the data management.
On the first screen I request the cocktail list. When a cocktail is selected I request the cocktail id endpoint.

B) Explain which library was used for the routing and why. Would you use the same for a consumer facing app targeting thousands of users? Why?
I'm using the navigation component provide from android

C) Have you used any strategy to optimize the performance of the list generated for the first feature?
No, but an optimization that could be done here is pagination if the API provides this feature.

D) Would you like to add any further comments or observations?
Yes, another optimitazion to the app could be done be spliting the api data model and the app data model. The api data model has a different fields for each ingredient and measure, so, those could be represented by a Ingredients object that containd the list of ingredient whit the measure.